Make view tests Qt5 only, re #24531
