+++ qt/python/CMakeLists.txt
@@ -91,7 +91,6 @@ endif ()
     mantidqt/widgets/codeeditor/test/test_execution.py
     mantidqt/widgets/codeeditor/test/test_interpreter.py
     mantidqt/widgets/codeeditor/test/test_multifileinterpreter.py
-    mantidqt/widgets/codeeditor/test/test_multifileinterpreter_view.py
     mantidqt/widgets/workspacewidget/test/test_workspacetreewidget.py
 
     mantidqt/widgets/test/test_jupyterconsole.py
@@ -125,6 +124,9 @@ endif ()
 
   # ctest target for widgets that only get tested in qt5, because they are only used in the workbench
   set ( PYTHON_WIDGET_QT5_ONLY_TESTS
+    mantidqt/widgets/codeeditor/test/test_interpreter_view.py
+    mantidqt/widgets/codeeditor/test/test_multifileinterpreter_view.py
+
     mantidqt/widgets/instrumentview/test/test_instrumentview_io.py
     mantidqt/widgets/instrumentview/test/test_instrumentview_view.py
   )
