+++ qt/python/mantidqt/widgets/codeeditor/test/test_interpreter.py
@@ -1,6 +1,6 @@
 # Mantid Repository : https://github.com/mantidproject/mantid
 #
-# Copyright &copy; 2017 ISIS Rutherford Appleton Laboratory UKRI,
+# Copyright &copy; 2019 ISIS Rutherford Appleton Laboratory UKRI,
 #     NScD Oak Ridge National Laboratory, European Spallation Source
 #     & Institut Laue - Langevin
 # SPDX - License - Identifier: GPL - 3.0 +
@@ -62,20 +62,6 @@ class PythonFileInterpreterTest(GuiTest):
             else:
                 self.assertEqual(w.clear_key_binding(key_combo), None,
                                  msg=fail_msg)
-
-    def test_find_replace_dialog(self):
-        w = PythonFileInterpreter()
-
-        # dialog not initialised by default, only when used
-        self.assertIsNone(w.find_replace_dialog)
-
-        w.show_find_replace_dialog()
-        self.assertIsNotNone(w.find_replace_dialog)
-
-        w.hide_find_replace_dialog()
-        # dialog not deleted on hide - just hidden
-        self.assertIsNotNone(w.find_replace_dialog)
-        self.assertFalse(w.find_replace_dialog.visible)
 
 
 if __name__ == '__main__':
